[#102] Rename common.js
